requirements.txt
django
djangorestframework                                                                                              